backend/requirements.txt
pytesseract
langchain_mistralai
openai
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
langchain-mistralai
chromadb
sentence-transformers
Pillow
pydub
SpeechRecognition 
SQLAlchemy 
bs4
requests
psycopg2
langchain_openai 
